+++ ole-app/olefs/src/main/java/org/kuali/ole/oleng/service/impl/OrderImportServiceImpl.java
@@ -1,14 +1,18 @@
 package org.kuali.ole.oleng.service.impl;
 
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.collections.map.HashedMap;
+import org.codehaus.jettison.json.JSONObject;
 import org.kuali.ole.DocumentUniqueIDPrefix;
 import org.kuali.ole.docstore.engine.service.storage.rdbms.pojo.BibRecord;
 import org.kuali.ole.constants.OleNGConstants;
+import org.kuali.ole.oleng.batch.process.model.ValueByPriority;
 import org.kuali.ole.oleng.batch.profile.model.BatchProcessProfile;
 import org.kuali.ole.oleng.batch.profile.model.BatchProfileDataMapping;
 import org.kuali.ole.oleng.resolvers.*;
 import org.kuali.ole.oleng.service.OrderImportService;
 import org.kuali.ole.pojo.OleTxRecord;
+import org.kuali.ole.spring.batch.BatchUtil;
 import org.kuali.ole.utility.MarcRecordUtil;
 import org.kuali.rice.krad.service.BusinessObjectService;
 import org.kuali.rice.krad.service.KRADServiceLocator;
@@ -17,6 +21,7 @@ import org.marc4j.marc.Record;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Created by SheikS on 1/6/2016.
@@ -38,14 +43,29 @@ public class OrderImportServiceImpl implements OrderImportService {
             List<Record> records = getMarcRecordUtil().convertMarcXmlContentToMarcRecord(bibRecord.getContent());
             Record marcRecord = records.get(0);
 
+            BatchUtil batchUtil = new BatchUtil();
+            batchUtil.sortDataMappings(batchProfileDataMappingList);
+            Map<String, List<ValueByPriority>> valueByPriorityMap = new HashedMap();
+
             for (Iterator<BatchProfileDataMapping> iterator = batchProfileDataMappingList.iterator(); iterator.hasNext(); ) {
                 BatchProfileDataMapping batchProfileDataMapping = iterator.next();
-
                 String destinationField = batchProfileDataMapping.getField();
+                boolean multiValue = batchProfileDataMapping.isMultiValue();
+                List<String> fieldValues = batchUtil.getFieldValues(marcRecord, batchProfileDataMapping, multiValue);
+
+                if (CollectionUtils.isNotEmpty(fieldValues)) {
+                    int priority = batchProfileDataMapping.getPriority();
+
+                    batchUtil.buildingValuesForDestinationBasedOnPriority(valueByPriorityMap, destinationField, multiValue, fieldValues, priority);
+                }
+            }
+
+            for (Iterator<String> iterator = valueByPriorityMap.keySet().iterator(); iterator.hasNext(); ) {
+                String destinationField =  iterator.next();
                 for (Iterator<TxValueResolver> valueResolverIterator = getValueResolvers().iterator(); valueResolverIterator.hasNext(); ) {
                     TxValueResolver txValueResolver = valueResolverIterator.next();
                     if (txValueResolver.isInterested(destinationField)) {
-                        String destinationValue = getDestinationValue(marcRecord, batchProfileDataMapping);
+                        String destinationValue = getDestinationValue(valueByPriorityMap,destinationField);
                         txValueResolver.setAttributeValue(oleTxRecord, destinationValue);
                     }
                 }
@@ -54,23 +74,18 @@ public class OrderImportServiceImpl implements OrderImportService {
         return oleTxRecord;
     }
 
-    private String getDestinationValue(Record marcRecord, BatchProfileDataMapping batchProfileDataMapping) {
-        String destValue = null;
-        if (batchProfileDataMapping.getDataType().equalsIgnoreCase(OleNGConstants.BIB_MARC)) {
-            String dataField = batchProfileDataMapping.getDataField();
-            String subField = batchProfileDataMapping.getSubField();
+    private String getDestinationValue(Map<String, List<ValueByPriority>> valueByPriorityMap, String fieldType) {
 
-            if (getMarcRecordUtil().isControlField(dataField)) {
-                destValue = getMarcRecordUtil().getControlFieldValue(marcRecord, dataField);
-            } else {
-                destValue = getMarcRecordUtil().getDataFieldValue(marcRecord, dataField, subField);
+        List<ValueByPriority> vals = valueByPriorityMap.get(fieldType);
+        for (Iterator<ValueByPriority> iterator1 = vals.iterator(); iterator1.hasNext(); ) {
+            ValueByPriority valueByPriority = iterator1.next();
+            List<String> values = valueByPriority.getValues();
+            if (CollectionUtils.isNotEmpty(values)) {
+                return values.get(0);
             }
-
-        } else if (batchProfileDataMapping.getDataType().equalsIgnoreCase(OleNGConstants.CONSTANT)) {
-            destValue = batchProfileDataMapping.getConstant();
         }
 
-        return destValue;
+        return null;
     }
 
     public MarcRecordUtil getMarcRecordUtil() {
