+++ ole-app/olefs/src/main/java/org/kuali/ole/spring/batch/BatchUtil.java
@@ -1,9 +1,16 @@
 package org.kuali.ole.spring.batch;
 
+import org.apache.commons.lang3.StringUtils;
 import org.kuali.incubator.SolrRequestReponseHandler;
+import org.kuali.ole.constants.OleNGConstants;
+import org.kuali.ole.oleng.batch.process.model.ValueByPriority;
+import org.kuali.ole.oleng.batch.profile.model.BatchProfileDataMapping;
 import org.kuali.ole.oleng.util.OleNgUtil;
 import org.kuali.ole.utility.MarcRecordUtil;
 import org.kuali.ole.utility.OleDsNgRestClient;
+import org.marc4j.marc.Record;
+
+import java.util.*;
 
 /**
  * Created by SheikS on 12/8/2015.
@@ -45,4 +52,95 @@ public class BatchUtil extends OleNgUtil{
     public void setMarcRecordUtil(MarcRecordUtil marcRecordUtil) {
         this.marcRecordUtil = marcRecordUtil;
     }
+
+    public void sortDataMappings(List<BatchProfileDataMapping> filteredDataMappings) {
+        Collections.sort(filteredDataMappings, new Comparator<BatchProfileDataMapping>() {
+            public int compare(BatchProfileDataMapping dataMapping1, BatchProfileDataMapping dataMapping2) {
+                int priorityForDataMapping1 = dataMapping1.getPriority();
+                int priorityForDataMapping2 = dataMapping2.getPriority();
+                return new Integer(priorityForDataMapping1).compareTo(new Integer(priorityForDataMapping2));
+            }
+        });
+    }
+
+
+    public List<String> getFieldValues(Record marcRecord, BatchProfileDataMapping batchProfileDataMapping, boolean multiValue) {
+        List<String> marcValues = new ArrayList<>();
+        if (batchProfileDataMapping.getDataType().equalsIgnoreCase(OleNGConstants.BIB_MARC)) {
+            String marcValue;
+            String dataField = batchProfileDataMapping.getDataField();
+            if (StringUtils.isNotBlank(dataField)) {
+                if (getMarcRecordUtil().isControlField(dataField)) {
+                    marcValue = getMarcRecordUtil().getControlFieldValue(marcRecord, dataField);
+                    marcValues.add(marcValue);
+                } else {
+                    String subField = batchProfileDataMapping.getSubField();
+                    if (multiValue) {
+                        marcValues = getMarcRecordUtil().getMultiDataFieldValues(marcRecord, dataField, batchProfileDataMapping.getInd1(), batchProfileDataMapping.getInd2(), subField);
+                    } else {
+                        marcValue = getMarcRecordUtil().getDataFieldValueWithIndicators(marcRecord, dataField, batchProfileDataMapping.getInd1(), batchProfileDataMapping.getInd2(), subField);
+                        if (StringUtils.isNotBlank(marcValue)) {
+                            marcValues.add(marcValue);
+                        }
+                    }
+                }
+            }
+        } else {
+            String constantValue = batchProfileDataMapping.getConstant();
+            if (StringUtils.isNotBlank(constantValue)) {
+                marcValues.add(constantValue);
+            }
+        }
+        return marcValues;
+    }
+
+    /**
+     * @param valueByPriorityMap
+     * @param destinationField
+     * @param multiValue
+     * @param fieldValues
+     * @param priority
+     * @return The valueByPriorityMap contains valueByPriority pojo for each destination field.
+     */
+    public Map<String, List<ValueByPriority>> buildingValuesForDestinationBasedOnPriority(Map<String, List<ValueByPriority>> valueByPriorityMap, String destinationField, boolean multiValue, List<String> fieldValues, int priority) {
+        List<ValueByPriority> valueByPriorities;
+
+        ValueByPriority valueByPriority = new ValueByPriority();
+        valueByPriority.setField(destinationField);
+        valueByPriority.setPriority(priority);
+        valueByPriority.setMultiValue(multiValue);
+
+        valueByPriority.setValues(fieldValues);
+
+        if (valueByPriorityMap.containsKey(destinationField)) {
+            valueByPriorities = valueByPriorityMap.get(destinationField);
+
+            if (valueByPriorities.contains(valueByPriority)) {
+                ValueByPriority existingValuePriority = valueByPriorities.get(valueByPriorities.indexOf(valueByPriority));
+                List<String> values = existingValuePriority.getValues();
+                values.addAll(fieldValues);
+                StringBuilder stringBuilder = new StringBuilder();
+                if (!multiValue) {
+                    for (Iterator<String> iterator = values.iterator(); iterator.hasNext(); ) {
+                        String value = iterator.next();
+                        stringBuilder.append(value);
+                        if (iterator.hasNext()) {
+                            stringBuilder.append(" ");
+                        }
+                    }
+                    values.clear();
+                    values.add(stringBuilder.toString());
+                }
+            } else {
+                valueByPriorities.add(valueByPriority);
+            }
+        } else {
+            valueByPriorities = new ArrayList<>();
+            valueByPriorities.add(valueByPriority);
+        }
+        valueByPriorityMap.put(destinationField, valueByPriorities);
+
+        return valueByPriorityMap;
+    }
+
 }
